requirements.txt
pyTelegramBotAPI
psycopg2-binary
flask
asyncpg
pytz
python-telegram-bot==20.7
